perf: add catMedium 64-bit inline bitset for categorical features
Production CPU profiling shows findInBitset consuming 59% of flat CPU
time due to indirect memory loads through catBoundaries/catThresholds
arrays. For categorical features with 33-64 categories, pack the bitset
into the Threshold float64 field directly (via Float64frombits), avoiding
the two-level indirection entirely.

Changes:
- Widen lgNode.Flags from uint8 to uint16 (struct size unchanged at
  24 bytes due to existing padding)
- Add catMedium flag (bit 8) for 64-bit inline bitsets
- Add catMedium fast path in categoricalDecision using math.Float64bits
- Detect bitsetSize==2 during model loading (both text and JSON formats)
  and pack into catMedium nodes

Made-with: Cursor

+++ skensemble_io.go
@@ -46,8 +46,8 @@ func lgTreeFromSklearnDecisionTreeRegressor(tree pickle.SklearnDecisionTreeRegre
 	createNode := func(idx int) (lgNode, error) {
 		node := lgNode{}
 		refNode := &tree.Tree.Nodes[idx]
-		missingType := uint8(0)
-		defaultType := uint8(0)
+		missingType := uint16(0)
+		defaultType := uint16(0)
 		node = numericalNode(uint32(refNode.Feature), missingType, refNode.Threshold, defaultType)
 		if tree.Tree.Nodes[refNode.LeftChild].LeftChild < 0 {
 			node.Flags |= leftLeaf
